see if test-ids error is because of abstract test class.
